+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added human modelling classes SWKHmiWorkbench and SWKHumanCatalog. Please note these are not documented in the
   CATIA V5 Basic Help file. Example of usage also added.
+* Added new class method Application.input_box(). @ptm-tm
 
 ## 0.6.0